TOT-133 : [REFACTOR] 계획 상세 정보 조회 DTO 통합
- PlanResponseDto로 통합.
- 수정페이지에서 사용될 계획 상세 정보 필드 추가 :
    imageThumbnail(사용자가 추가했던 장소의 썸네일),
    regionId(),
    placeName(사용자가 추가했던 장소의 이름),
    contentTypeId(사용자가 추가했던 장소의 타입),
    subCategoryName(사용자가 추가했던 장소의 카테고리명),
    latitude, longitude(사용자가 추가했던 장소를 지도에 표시하기 위한 위도, 경도)
- 지역정보를 조회하기 위한 필드 추가 :
    RegionResponseDto(지역아이디, 지역이름)
